python-processor/requirements.txt
# python-processor/requirements.txt
google-cloud-storage
google-cloud-documentai
google-generativeai
pypdf
python-dotenv
